# avansysTest/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Usuários",
        "description": "API para gerenciamento de usuários",
        "contact": {
            "email": "suporte@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/post": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Listar todos os posts",
                "description": "Display a listing of the resource.",
                "operationId": "425e95aaa6ab243fc407ffcf09db9de5",
                "responses": {
                    "200": {
                        "description": "Lista de posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Criar um novo post",
                "description": "Store a newly created resource in storage.",
                "operationId": "35cc3a4123556e20dc17f6f45be00f18",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/post/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Obter um post pelo ID",
                "description": "Display the specified resource.",
                "operationId": "820a95132e64f02c9c71b2cfc3cadb0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do post"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Post"
                ],
                "summary": "Atualizar um post",
                "description": "Update the specified resource in storage.",
                "operationId": "d80bce236721ddcb1bdc86a4e869be81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post atualizado com sucesso"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Deletar um post",
                "description": "Remove the specified resource from storage.",
                "operationId": "cdf3a5c9a54637677e9d19e0bd933ab4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post deletado com sucesso"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            }
        },
        "/tag": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Listar todas as tags",
                "description": "Display a listing of the resource.",
                "operationId": "8646a95de680134a63803fc1ee2f7fd9",
                "responses": {
                    "200": {
                        "description": "Lista de tags",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tag"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tag"
                ],
                "summary": "Criar uma nova tag",
                "description": "Store a newly created resource in storage.",
                "operationId": "6be3be9f91a802348cbc34de1d09626a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TagRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tag criada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/tag/{id}": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Obter uma tag pelo ID",
                "description": "Display the specified resource.",
                "operationId": "02156ea53e4153b55e317eaf69b0915d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da tag"
                    },
                    "404": {
                        "description": "Tag não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tag"
                ],
                "summary": "Atualizar uma tag",
                "description": "Update the specified resource in storage.",
                "operationId": "3f481624cd7a71a740113c5d9ffda942",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TagRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tag atualizada com sucesso"
                    },
                    "404": {
                        "description": "Tag não encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tag"
                ],
                "summary": "Deletar uma tag",
                "description": "Remove the specified resource from storage.",
                "operationId": "460442a8e0259c45a24b940b790db942",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tag deletada com sucesso"
                    },
                    "404": {
                        "description": "Tag não encontrada"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Listar todos os usuários",
                "description": "Display a listing of the resource.",
                "operationId": "23679253d16f34d2972249751e3631dd",
                "responses": {
                    "200": {
                        "description": "Lista de usuários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Criar um novo usuário",
                "description": "Store a newly created resource in storage.",
                "operationId": "ce3cf31668a69eb39e5e98738715afa1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso"
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Obter um usuário pelo ID",
                "description": "Display the specified resource.",
                "operationId": "1ec60d4ee968175164a94c21f897f8ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do usuário"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Atualizar um usuário",
                "description": "Update the specified resource in storage.",
                "operationId": "786a794071cb7098f55a4a4911e7d3f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado com sucesso"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Deletar um usuário",
                "description": "Remove the specified resource from storage.",
                "operationId": "a56b12a3694a926699892056cdcf8a9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuário deletado com sucesso"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "date_birth": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "UserRequest": {
                "required": [
                    "name",
                    "email",
                    "date_birth"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "date_birth": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "Tag": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "TagRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Post": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "PostRequest": {
                "required": [
                    "title",
                    "content",
                    "user_id"
                ],
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Post",
            "description": "Operações relacionadas a posts"
        },
        {
            "name": "Tag",
            "description": "Operações relacionadas a tags"
        },
        {
            "name": "User",
            "description": "Operações relacionadas a usuários"
        }
    ]
}